Improve BPF tracepoint probe derivation
One goal of BPF tracepoint probe derivation is to determine the offsets of the members
of the BPF tracepoint context structs. These members are the probe's context variables.
The context structs always contain 8 bytes of padding before their first member.

During the probe derivation process, the context structs are replicated in the tracequery
modules and we read the module's debuginfo in order to determine the offsets of the
struct's members. Previously, the 8 bytes of padding was not included in the tracequery
structs and instead 8 was added to the offsets of members of these structs. This approach
ignores the possibility that these 8 bytes of padding could influence the amount of
padding between the struct's other members. This would result in incorrect offsets.

This patch adds an 8 byte member to the front of each of the tracequery structs to ensure
they are proper replicas of the tracepoint context structs.