BPF translator: array elements behave as if preinitialized to 0
Previously we exited a probe if there was an attempt to use an element
that an array did not contain. This does not match the behavior of the
default runtime which acts as though the element was preinitialized to
0. This commit changes the translator to match this behavior.

bpf-translate.cxx (visit_arrayindex): If map_lookup_elem returns 0,
set result to 0 instead of jumping to exit_block.

bpf_tests/array_preinit.stp: tests this change.