Fix the systemtap testsuite LD_LIBRARY_PATH.
The goal here is to ensure the right dyninst libraries are used.
Systemtap is not only packaged for base RHEL/CentOS, but also as
a part of SCL (red hat developer toolset).  In case of SCL we need
to ensure that SCL version of systemtap tests against SCL version
of dyninst and the same is true for base os versions of both
components.

Before this change devtoolset version of systemtap was running
its tests against base rhel version of dyninst, and if that was
missing, dyninst tests were crashing.

This is a testsuite problem only, unrelated to normal stap usage.