PR23160,PR14690: support /* guru */ embedded-C for CONTEXT->sregs
For the newfangled syscall support, an embedded-C function is needed
to set the CONTEXT->sregs field from any of the new gajillion syscall
probe aliases.  Since we want that ->sregs field trustworthy, and not
callable by the user with some random junk pointer, we want this
function to be only callable from the tapset, but not replicated (so
not "private" in each file).

To do this, we extend the parser/elaborator to extend the logic of /*
guru */ markup in embedded-C functions.  Namely, calls to such
functions from the tapset (which is parsed with privileged flags) are
now permitted, kind of as if they were a private copy in each tapset
file that called them.  The stapfile privileged field, which has
fallen into disuse, is brought back to life and propagates the
parse-time pf_guru flag.  It's used with a new visitor that checks
each caller-callee relationship in function calls, moving some similar
code from staptree.cxx.

The syscall.read aliases are adapted to use the new single copy of
__set_syscall_pt_regs() in aux2_syscalls.stp.