PR23356: stap-serverd: switch back to dbm: nss databases
The sqlite sql: nss ones seem to have different
authentication/login/password protocols as on rawhide, and our logic
can't quite play right with them.  Maybe it's due to possible
concurrent access from a stap-serverd and stap client (when both run
as root).  Whatever it is, plain dbm: seems to restore function,
even if with a warning about "legacy database".