stapbpf: add sprintf support to user space interpreter.
This patch implements sprintf for probes that run in stapbpf's user
space BPF interpreter. String support is still very limited. The
return value of sprintf can be assigned to a local variable and
passed to printf as an argument, but not much else. The purpose
of this patch is to add just enough string support for stapbpf
procfs probes to be useful (see PR23285).

* bpf-internal.h: add bpf_func_id for sprintf.

* bpf-translate.cxx (visit_print_format): Add logic for returning
  the string instead of calling trace_printk.

* bpfinterp.cxx (bpf_interpret, bpf_sprintf): Add handler for sprintf
  call to interpreter.

* bpf_tests/sprintf.stp: tests these changes.