Add the also_ran.stp script to the stap-export examples
The also_ran.stp script provides a tally of the executables and shared
libraries run on the system.  The counts provide some indication of
how frequently particular executables and shared libraries are
used. The executables and share libraries could be mapped back to the
packages to give an indication of what software packages are being
used on the system.