Allow more flexible Prometheus output formatting
There are cases where would like adjust the output of the data being
generated in Prometheus format.  For example, storing syscall numbers
to minimize the overhead of recording the information and then map
syscall numbers to more symbolic names when generating the Prometheus
formated data.

The existing prometheus_dump_array* macros work as before and there is
now a matching set of prometheus_dump_array_map* macros that have
additional arguments to pass in mapping functions.  For unmodified
fields the sprint function is used.  Below is a use where the "count"
and "pid" fields are printed out as the default
@prometheus_dump_array2 would print them and the "syscall" field is
translated from a number to the syscall name by the syscall_name
function.

    @prometheus_dump_array2_map(arr, "count", "pid", "syscall",
        sprint, sprint, syscall_name )