Use the flexible Prometheus formatting to lower cost of recording data
We want to keep the recording code as simple as possible to reduce the
overhead.  Recording the syscall number avoids making a function call,
generating a string for each syscall, and has simpler associative
array indexing.

The one down side of this approach is on 64-bit machines where there
are differences between 32-bit and 64-bit syscall numbering and names
32-bit code is going to get the wrong 64-bit syscall names for the
32-bit syscalls.