Add support for return statements without values
For user-defined stap functions which do not return any values, it can
be convenient to allow *bare* return statements to take shortcuts in
the control flow of the function body.

The parser treats a following semicolon (';') or a closing curly bracket
('}') as a terminator for such bare return statements.

Added tests to cover various cases like use of plain 'return' in a
function actually returning some values. Also added tests to make sure
the pretty-printer adds a trailing semicolon for such bare return
stateuments for various arrangements. Both the "kernel" and "dyninst"
runtimes are covered.

A fix for the bpf translator is also included.

Updated NEWS to document this new feature.