Make the 3rd operand of ternary '?:' bind tighter than binary '='
In the C language, the 3rd operand of the ternary operator binds tighter
than the binary assignment operators. It is better for the stap language
to be consistent with C in operator precedence.

Added several tests to check the precedence of the ternary operator,
including nested ternary operator expressions. I've verified the results
with similar C programs with gcc myself.

Fixed an existing test case under systemtap.examples/, ansi_colors2.stp,
which incorrectly assumed that `+=` binds tighter than the 3rd operand
of the ternary operator.

The original behavior can be restored by the --compatible 3.3 option.

Updated NEWS to reflect this backward-incompatible change in the parser.