Have whythefail.stp probe a function that's name has not changed
The Linux 4.17 kernel has made a number of changes in the syscall
function names.  These changes caused the whythefail.stp test to fail.
Rather than probing the sys_open function which no longer exists in
the 4.17 kernels the tests are now using the do_sys_open function
which remains the same.