PR23160,PR14690: fix more undeclared macro errors
Moved contents of aux2_syscall.stp to aux_syscall.stp. Since the
contents of aux2_syscall.stp are going to be used by all syscalls,
doing this will cause aux_syscall.stp to always be included and so
will the header files containing the macros that are #included in
aux_syscall.stp.