Adjust the periodic.stp example to work with newer Linux kernels
The data field in the timer_list struct was removed in newer kernels.
The various functions executed when a timer expires now use
container_of macros to find the struct that the timer_list was
embedded in.  The periodic.stp script has been modified to use
container_of when the data field is not available.