Update pfiles.stp to work with Linux 4.17 and newer
The ABI for getname was changed in Linux 4.17 by commit 9b2c45d479d to
eliminate passing in a pointer to the store length information.  For
the newer version of the getname function negative return values are
errors and non-negative return value are the length.  The changes to
the pfiles.stp script picks the appropriate ABI based on the kernel
version.