Adjust noptrace.stp to avoid modifying ptrace syscall arguments
SystemTap has mutiple mechanisms to instrument the syscalls. For newer
kernels the non-dwarf versions are more likely to be used than the
dwarf-based ones.  Target variables are not available in the non-dwarf
version, making it impossible to change the value of a target variable
$request in syscall.ptrace.  An alternative way of forcing the ptrace
calls to fail is now used in noptrace.stp.  The capability checks are
instrumented.  When the ptrace syscall calls the capability check, the
capability checks are forced to return -EPERM, causing the ptrace
syscalls to fail.