Match arm64 non-dwarf syscall probe points
The x86_64 linux 4.17 prefixes the syscall function names. The
systemtap tapsets use the arch_syscall_prefix macro to get the proper
prefix.  The arm64 kernel also has prefixes for the syscall function
names, but the arch_syscall_prefix needs to be __arm64_ not __x86_64_
or __ia32_.  Adjusted syscalls.stpm to pick the appropriate prefix for
the architecture.  This patch makes it easier to do the same for any
other architectures that prefix the syscall names.