Add more __NR_* for missing defines on aarch64 (and ppc64)
There are a number of syscall defines (__NR_*) on the x86_64 that are
not on aarch64 (or ppc64).  They need to have something defined for
those undefine constants, so the kernel module code for the fallback
tp_syscall.* and tp_syscall.*.return probes compile and do not throw
"'__NR_blah' undeclared" errors.

Most of the added __NR_* defines are for aarch64, but the ppc64 also
needs the __NR_compat_bdflush.