Do the search save_stack_trace_regs() in the stap module's initialization
Previously, the search for save_stack_trace_regs() was done the first
time the _stp_stack_print_fallback was called. However, want to avoid
calling kallsyms_lookup_name() in some arbitrary probe handler as it
may take a signficant amount of time to search for
save_stack_trace_regs().  Moving this search to where other
initialization operations are done and there are fewer contraints on
runninging it.