Make varwatch.stp work regardless of kprocess.release and a bit more flexible
With newer kernels the kprocess.release is inlined in some places and
the desired argument $p is unavailable.  Make the script just over
writing over older entries in the var associative array.

The script would only work for variables that resulted in strings.
Someone may want to use this script to monitor a single variable that
is an integer.  Having the probe never force the associative array
store strings will cause a problem for monitoring a single integer
variable.  Removed the probe never as the normal use cases of this
script will have something set the type of the associative array.  If
it doesn't, there aren't any value changes to monitor.